move supported locals relative to the scripts.
