+++ scripts/aggregate-translations/i18nSupportedLocales.js
